Split valentines editor into separate page
