+++ src/pages/modules/valentines/index.tsx
@@ -1,6 +1,7 @@
 import { Route, Routes } from "react-router-dom";
 import Valentines from "./Valentines";
 import LoginGate from "../../../comps/ui/LoginGate";
+import Create from "./Create";
 
 // overkill but leaving room for growth if needed
 const ValentinesRouter = () => {
@@ -8,6 +9,7 @@ const ValentinesRouter = () => {
         <LoginGate>
             <Routes>
                 <Route path="/" Component={Valentines} />
+                <Route path="/create" Component={Create} />
             </Routes>
         </LoginGate>
     );
